init all of the file & requirements for quantization & fine-tuning
