+++ requirements.txt
@@ -1,22 +1,22 @@
-aws-requests-auth==0.4
-boto3==1.28
-botocore==1.31
-fastapi>=0.104,<0.105
-gunicorn==21.2.0
-httpx>=0.23.3
-pydantic==2.4.2
-pydantic-core==2.10.1
-pytest==7.3.2
-python-dotenv>=1.0.0
-requests-auth-aws-sigv4==0.7
-torch>=2.1,<2.2
-transformers[sentencepiece]==4.35
-uvicorn[standard]==0.24
-openai>=1.7.1,<1.8
-trl==0.7.1
-bitsandbytes==0.41.1 
-datasets==2.14.4
-peft==0.5.0
-fsspec==2023.9.2 
-scikit-learn==1.4.0
-accelerate==0.32.1
+aws-requests-auth
+boto3
+botocore
+fastapi
+gunicorn
+httpx
+pydantic
+pydantic-core
+pytest
+python-dotenv
+requests-auth-aws-sigv4
+torch
+transformers[sentencepiece]
+uvicorn[standard]
+openai
+trl
+bitsandbytes
+datasets
+peft
+fsspec
+scikit-learn
+accelerate
