+++ training/config/parameters.py
@@ -1,13 +1,9 @@
 import torch
-from transformers import TrainingArguments
-from transformers import BitsAndBytesConfig
-from peft import PeftModel
-from transformers import AutoModelForCausalLM, AutoTokenizer
-import torch
-from fine_tuning_configs.general_config import base_model,save_dir
-import re
+from transformers import (
+    TrainingArguments, 
+    BitsAndBytesConfig, 
+)
 from peft import LoraConfig
-from transformers import AutoModelForCausalLM
 
 # Type to cast the models submodules
 compute_dtype = getattr(torch, "float16")
@@ -22,41 +18,6 @@ quant_config = BitsAndBytesConfig(
     # which is defined by the module that includes the submodules
     bnb_4bit_use_double_quant=False,  # controls whether we use a second quantization to save an additional 0.4 bits per parameter
 )
-
-
-def merge_and_save(save_dir:str=save_dir,base_model:str=base_model):
-    """
-    Will merge the fine-tuned weights to the original model and save it to the chosen dir
-
-    Parameters
-    ----------
-    save_dir : str
-        fine-tuned model save directory
-    
-    base_model : str
-        HF identifier of the base model
-
-    Returns
-    -------
-    A saved & merged fine-tuned model in the selected local directory
-    """
-    # Reload model in FP16 and merge it with LoRA weights
-    base_model = AutoModelForCausalLM.from_pretrained(
-                base_model,
-                low_cpu_mem_usage=True,
-                return_dict=True,
-                torch_dtype=torch.float16,
-                device_map={"": 0},
-            )
-    try:
-        model = PeftModel.from_pretrained(base_model, save_dir)
-        merged_model = model.merge_and_unload()
-        merged_model.save_pretrained(save_dir, safe_serialization=True)
-        return "Merge done"
-    except Exception as e:
-        return f"An unexpected error as occured: {e}"
-
-
 
 lora_alpha = (
     16,
@@ -81,73 +42,13 @@ bias = (
 
 task_type = "CAUSAL_LM"
 
-
-def get_target_modules(model: AutoModelForCausalLM) -> list:
-    """
-    Search the llm architecture for all of the linear layers, so that we can update them during fine-tuning
-
-    Parameters
-    ----------
-    model : AutoModelForCausalLM
-        The LLM
-
-    Returns
-    -------
-    target_modules : list
-        list of all of the linear layers in the LLM architecture
-    """
-    model_modules = str(model.modules)
-    pattern = r"\((\w+)\): Linear"
-    linear_layer_names = re.findall(pattern, model_modules)
-
-    names = []
-    # Print the names of the Linear layers
-    for name in linear_layer_names:
-        names.append(name)
-    target_modules = list(set(names))
-    return target_modules
-
-
-def get_qlora_config(
-    model: AutoModelForCausalLM,
-    lora_alpha: int = lora_alpha,
-    lora_dropout: float = lora_dropout,
-    r: int = r,
-    bias: str = bias,
-    task_type: str = task_type,
-) -> LoraConfig:
-    """Config for the QLoRA fine-tuning
-
-    Parameters
-    ----------
-    model : AutoModelForCausalLM
-        LLM
-    target_modules : list
-        All of the linear layers prensent in the LLM
-    lora_alpha : int, optional
-        _description_, by default lora_alpha
-    lora_dropout : float, optional
-        standard dropout, by default lora_dropout
-    r : int, optional
-        rank of the low rank matrices, by default r
-    bias : str, optional
-        _description_, by default bias
-    task_type : str, optional
-        type of task that we will be fine-tuning for, by default task_type
-
-    Returns
-    -------
-    LoraConfig
-        _description_
-    """
-    return LoraConfig(
-        lora_alpha=lora_alpha,
-        lora_dropout=lora_dropout,
-        r=r,
-        bias=bias,
-        task_type=task_type,
-        target_modules=get_target_modules(model),
-    )
+peft_params = LoraConfig(
+    lora_alpha=lora_alpha,
+    lora_dropout=lora_dropout,
+    r=r,
+    bias=bias,
+    task_type=task_type
+)
 
 
 # Set training parameters
@@ -174,3 +75,4 @@ training_params = TrainingArguments(
     lr_scheduler_type="cosine"  # "constant",
     # report_to="tensorboard"
 )
+
